build files: Use Ceylon 1.1.0, ceylon.test 1.1.0
This assumes a setup where you have a Ceylon 1.1.0 installation
(release) in your ~/bin/ceylon-1.1.0, and a ceylon-release executable
linked to that.

This setup can be reproduced with the following commands:

    cd ~/bin;
    wget http://downloads.ceylon-lang.org/cli/ceylon-1.1.0.zip;
    unzip -q ceylon-1.1.0.zip;
    rm ceylon-1.1.0.zip;
    ln -s ceylon-1.1.0/bin/ceylon ceylon-release;

The test target now uses <exec> instead of <ceylon-test> because we need
to supply the ceylon.test version (it defaults to the latest available),
which is only supported by the CLI tool, not the ant task.