Parse base types with type arguments
No use-site variance yet.

Another significant change is that the parse methods all no longer
return an optional type: If we descend into that method, we have already
decided (through lookahead) that we want to parse using this
alternative, so we shouldn’t bother the caller with having to work
around the return value possibly being null – instead, the parse method
itself should do as well as it can, and then return a convincing fake.